requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
yfinance==0.2.31
requests==2.31.0
pandas==2.1.4
ta==0.11.0
python-telegram-bot==20.7
apscheduler==3.10.4
aiohttp==3.9.1
pydantic==2.5.2
